Add responsive default ui attribute to PageModel class.
